Call to prayer scene done
Finished call to prayer scene with UI prompt and also created the LoadNextScene script. The SceneTransitioner game object will remain between scene loads to handle scene shifting. In order to incorporate this, you only need to create a button, drag SceneTransitioner as an onClick event with the function "OnButtonClicked", and specify the next scene's name in the String input box.